Version 2.6.0: -replaced static fields by static properties and dependency properties
- improved class TileLayer
- MapGraticule with arc seconds

+++ MapControl/MapBase.Silverlight.WinRT.cs
@@ -20,31 +20,32 @@ namespace MapControl
     public partial class MapBase
     {
         public static readonly DependencyProperty ForegroundProperty = DependencyProperty.Register(
-            "Foreground", typeof(Brush), typeof(MapBase), new PropertyMetadata(new SolidColorBrush(Colors.Black)));
+            "Foreground", typeof(Brush), typeof(MapBase),
+            new PropertyMetadata(new SolidColorBrush(Colors.Black)));
 
         public static readonly DependencyProperty CenterProperty = DependencyProperty.Register(
-            "Center", typeof(Location), typeof(MapBase), new PropertyMetadata(new Location(),
-                (o, e) => ((MapBase)o).CenterPropertyChanged((Location)e.NewValue)));
+            "Center", typeof(Location), typeof(MapBase),
+            new PropertyMetadata(new Location(), (o, e) => ((MapBase)o).CenterPropertyChanged((Location)e.NewValue)));
 
         public static readonly DependencyProperty TargetCenterProperty = DependencyProperty.Register(
-            "TargetCenter", typeof(Location), typeof(MapBase), new PropertyMetadata(new Location(),
-                (o, e) => ((MapBase)o).TargetCenterPropertyChanged((Location)e.NewValue)));
+            "TargetCenter", typeof(Location), typeof(MapBase),
+            new PropertyMetadata(new Location(), (o, e) => ((MapBase)o).TargetCenterPropertyChanged((Location)e.NewValue)));
 
         public static readonly DependencyProperty ZoomLevelProperty = DependencyProperty.Register(
-            "ZoomLevel", typeof(double), typeof(MapBase), new PropertyMetadata(1d,
-                (o, e) => ((MapBase)o).ZoomLevelPropertyChanged((double)e.NewValue)));
+            "ZoomLevel", typeof(double), typeof(MapBase),
+            new PropertyMetadata(1d, (o, e) => ((MapBase)o).ZoomLevelPropertyChanged((double)e.NewValue)));
 
         public static readonly DependencyProperty TargetZoomLevelProperty = DependencyProperty.Register(
-            "TargetZoomLevel", typeof(double), typeof(MapBase), new PropertyMetadata(1d,
-                (o, e) => ((MapBase)o).TargetZoomLevelPropertyChanged((double)e.NewValue)));
+            "TargetZoomLevel", typeof(double), typeof(MapBase),
+            new PropertyMetadata(1d, (o, e) => ((MapBase)o).TargetZoomLevelPropertyChanged((double)e.NewValue)));
 
         public static readonly DependencyProperty HeadingProperty = DependencyProperty.Register(
-            "Heading", typeof(double), typeof(MapBase), new PropertyMetadata(0d,
-                (o, e) => ((MapBase)o).HeadingPropertyChanged((double)e.NewValue)));
+            "Heading", typeof(double), typeof(MapBase),
+            new PropertyMetadata(0d, (o, e) => ((MapBase)o).HeadingPropertyChanged((double)e.NewValue)));
 
         public static readonly DependencyProperty TargetHeadingProperty = DependencyProperty.Register(
-            "TargetHeading", typeof(double), typeof(MapBase), new PropertyMetadata(0d,
-                (o, e) => ((MapBase)o).TargetHeadingPropertyChanged((double)e.NewValue)));
+            "TargetHeading", typeof(double), typeof(MapBase),
+            new PropertyMetadata(0d, (o, e) => ((MapBase)o).TargetHeadingPropertyChanged((double)e.NewValue)));
 
         partial void Initialize()
         {
@@ -67,7 +68,7 @@ namespace MapControl
         private void SetViewportTransform(Location origin)
         {
             MapOrigin = mapTransform.Transform(origin);
-            ViewportScale = Math.Pow(2d, ZoomLevel) * TileSource.TileSize / 360d;
+            ViewportScale = Math.Pow(2d, ZoomLevel) * (double)TileSource.TileSize / 360d;
 
             viewportTransform.Matrix =
                 new Matrix(1d, 0d, 0d, 1d, -MapOrigin.X, -MapOrigin.Y)
